% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHosvd.R
\name{computeHosvd}
\alias{computeHosvd}
\title{Title
Compute higher order singular value decomposition}
\usage{
computeHosvd(Z, dims = c(10, dim(Z@value)[-1]), scale = T)
}
\arguments{
\item{Z}{array that includes omics data}

\item{dims}{dimensions to be compued by HOSVD}

\item{scale}{If value is scaled}
}
\value{
List that includes output from HOSVD
}
\description{
Title
Compute higher order singular value decomposition
}
\examples{
Z <- PrepareSummarizedExperimentTensor(sample=matrix(as.character(1:6),c(3,2)),
                                      feature=as.character(1:10),
                                      value=array(runif(10*3*2),c(10,3,2)))
HOSVD <- computeHosvd(Z)
}
