% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHosvdSqure.R
\name{computeHosvdSqure}
\alias{computeHosvdSqure}
\title{Title
Compute higher order singular value decomposition from the tensor generated from squared matrix}
\usage{
computeHosvdSqure(
  Z,
  dims = unlist(lapply(dim(Z@value), function(x) {
     min(10, x)
 })),
  scale = T
)
}
\arguments{
\item{Z}{A tensor including sample names, feature values, associated with featureRange and sample properties}

\item{dims}{dimensions to be compued by HOSVD}

\item{scale}{If value is scaled}
}
\value{
List that includes output from HOSVD
}
\description{
Title
Compute higher order singular value decomposition from the tensor generated from squared matrix
}
\examples{
omics1 <- matrix(runif(100),10)
dimnames(omics1) <- list(1:10,1:10)
omics2 <- matrix(runif(100),10)
dimnames(omics2) <- dimnames(omics1)
Multi <- list(omics1,omics2)
Z <- PrepareSummarizedExperimentTensorSquare(sample=matrix(colnames(omics1),1),
            feature=list(omics1=rownames(omics1),
            omics2=rownames(omics2)),
            value=convertSquare(Multi),
            sampleData=list(NA))
            HOSVD <- computeHosvdSqure(Z)
}
