% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSquare.R
\name{convertSquare}
\alias{convertSquare}
\title{Generate squared tensor from multiomics data}
\usage{
convertSquare(Multi)
}
\arguments{
\item{Multi}{A list that include multiomics data}
}
\value{
A tensor computed from multiomics data
}
\description{
Generate squared tensor from multiomics data
}
\examples{
omics1 <- matrix(runif(100),10)
dimnames(omics1) <- list(seq_len(10),seq_len(10))
omics2 <- matrix(runif(100),10)
dimnames(omics2) <- dimnames(omics1)
Multi <- list(omics1,omics2)
Z <- convertSquare(Multi)
}
