% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSingularValueVectorSmall.R
\name{selectSingularValueVectorSmall}
\alias{selectSingularValueVectorSmall}
\title{Title
Select singular value vectors from HOSVD}
\usage{
selectSingularValueVectorSmall(HOSVD, input_all = NULL)
}
\arguments{
\item{HOSVD}{output from HOSVD}

\item{input_all}{if ist is no null, no interactive mode is
activated but provided values are used.}
}
\value{
Selected singular value vector IDs
}
\description{
Title
Select singular value vectors from HOSVD
}
\examples{
Z <- PrepareSummarizedExperimentTensor(
sample=matrix(as.character(seq_len(6)),c(3,2)),
feature=as.character(seq_len(10)),
 value=array(runif(10*3*2),c(10,3,2)))
 HOSVD <- computeHosvd(Z)
input_all <- selectSingularValueVectorSmall(HOSVD,input_all=c(1,1))
}
