% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSingularValueVectorSmall.R
\name{selectSingularValueVectorSmall}
\alias{selectSingularValueVectorSmall}
\title{Title
Select singular value vectors from HOSVD}
\usage{
selectSingularValueVectorSmall(HOSVD)
}
\arguments{
\item{HOSVD}{output from HOSVD}
}
\value{
Selected singular value vector IDs
}
\description{
Title
Select singular value vectors from HOSVD
}
\examples{
Z <- PrepareSummarizedExperimentTensor(sample=matrix(as.character(1:6),c(3,2)),
                                      feature=as.character(1:10),
                                      value=array(runif(10*3*2),c(10,3,2)))
                                      HOSVD <- computeHosvd(Z)
              input_all <- selectSingularValueVectorSmall(HOSVD)
}
