% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDbasedUFE-package.R
\docType{package}
\name{TDbasedUFE-package}
\alias{TDbasedUFE}
\alias{TDbasedUFE-package}
\title{TDbasedUFE: Tensor Decomposition Bassed Unsupervised Feature Extraction}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tagtag/TDbasedUFE}
  \item Report bugs at \url{https://github.com/tagtag/TDbasedUFE/issues}
}

}
\keyword{internal}
